requirements.txt
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
gunicorn==20.1.0
Flask==2.2.2
python-dotenv==0.19.1
openai==0.25.0